modularize ai demo
